v1.0.5
Fixing problems observed in production: I thin kthey were due to a wrong way of passing the $parameters. Moreover in the Close button of the common setting panel there was still an old part of a code, asking for $Board, which is not actually passed to the panel because it's not needed. I've removed it.